+++ packages/manager/modules/iplb/src/dashboard/iplb-detail.controller.js
