+++ packages/manager/modules/iplb/src/dashboard/iplb.less
