+++ packages/manager/modules/iplb/src/dashboard/translations/Messages_fr_CA.json
@@ -68,23 +68,23 @@
   "iplb_cancel": "Annuler",
   "iplb_activate": "Activer",
   "iplb_deactivate": "Désactiver",
-  "iplb_edit": "Modifier",                                              
-  "iplb_consult": "Consulter",                                                    
-  "iplb_manage": "Gérer",                                                        
-  "iplb_modify": "Modifier",                                                  
-  "iplb_select": "Sélectionner",                                                  
-  "iplb_add": "Ajouter",                                                       
-  "iplb_delete": "Supprimer",                                                
-  "iplb_remove": "Retirer",                                                          
-  "iplb_modal_ok": "OK",                                                          
-  "iplb_tile_subscription": "Abonnement",                                         
-  "iplb_tile_subscription_offer": "Offre",                                        
-  "iplb_tile_subscription_contact": "Contacts",                                   
-  "iplb_tile_subscription_contact_admin": "Administrateur",                       
-  "iplb_tile_subscription_contact_billing": "Facturation",                        
-  "iplb_tile_subscription_contact_technical": "Technique",                        
-  "iplb_tile_subscription_creation_date": "Date de création",                     
-  "iplb_preview_see": "Voir l'aperçu",                                           
+  "iplb_edit": "Modifier",
+  "iplb_consult": "Consulter",
+  "iplb_manage": "Gérer",
+  "iplb_modify": "Modifier",
+  "iplb_select": "Sélectionner",
+  "iplb_add": "Ajouter",
+  "iplb_delete": "Supprimer",
+  "iplb_remove": "Retirer",
+  "iplb_modal_ok": "OK",
+  "iplb_tile_subscription": "Abonnement",
+  "iplb_tile_subscription_offer": "Offre",
+  "iplb_tile_subscription_contact": "Contacts",
+  "iplb_tile_subscription_contact_admin": "Administrateur",
+  "iplb_tile_subscription_contact_billing": "Facturation",
+  "iplb_tile_subscription_contact_technical": "Technique",
+  "iplb_tile_subscription_creation_date": "Date de création",
+  "iplb_preview_see": "Voir l'aperçu",
   "iplb_complete_order": "Régler le bon de commande",
-  "iplb_tile_subscription_autorenew": "Renouvellement automatique"   
+  "iplb_tile_subscription_autorenew": "Renouvellement automatique"
 }
