+++ packages/manager/modules/iplb/src/dashboard/translations/Messages_fr_FR.json
@@ -1,0 +1,90 @@
+{
+  "iplb_back_home": "Retour",
+  "iplb_guides": "Guides",
+  "iplb_guides_title": "IPLB",
+  "iplb_guide_footer": "Tous les guides OVH",
+  "iplb_tab_home": "Accueil",
+  "iplb_tab_frontends": "Frontends",
+  "iplb_tab_server_farm": "Fermes de serveurs",
+  "iplb_tab_ssl_certificate": "Certificats SSL",
+  "iplb_tab_vrack": "Réseaux privés",
+  "iplb_tab_graph": "Graphiques",
+  "iplb_tab_task": "Tâches",
+  "iplb_modal_failover_ip_title": "IP Failover",
+  "iplb_modal_nat_ip_title": "IPv4 de sortie",
+  "iplb_modal_nat_ip_description": "Vous devez autoriser ces adresses IP à joindre vos serveurs.",
+  "iplb_modal_name_change_title": "Modifier le nom",
+  "iplb_modal_name_change_service_name_label": "Nom par défaut",
+  "iplb_modal_name_change_display_name_label": "Nom personnalisé",
+  "iplb_modal_name_change_updating_error": "Erreur lors du changement de nom.",
+  "iplb_modal_cipher_change_title": "Modifier les ciphers",
+  "iplb_modal_cipher_change_loading_error": "Erreur lors du chargement de la modale de modification des ciphers.",
+  "iplb_modal_cipher_change_updating_error": "Erreur lors du changement de cipher.",
+  "iplb_modal_cipher_change_cipher_intermediate_title": "Intermédiaire",
+  "iplb_modal_cipher_change_cipher_intermediate_description": "Clients compatibles dès Android 2.3, IE 7, Java 7, Opera 5, Windows XP IE 8",
+  "iplb_modal_cipher_change_cipher_modern_title": "Moderne",
+  "iplb_modal_cipher_change_cipher_modern_description": "Sécurité accrue mais compatible avec moins de systèmes, Clients compatibles dès Android 5.0, Chrome 30, Edge, Firefox 27, IE 11 on Windows 7, Java 8, Opera 17, Safari 9",
+  "iplb_modal_offer_change_title": "Modifier l'offre",
+  "iplb_information_loading_error": "Erreur lors du chargement de la section Information.",
+  "iplb_configuration_loading_error": "Erreur lors du chargement de la section Configuration.",
+  "iplb_subscription_loading_error": "Erreur lors du chargement de la section Abonnement.",
+  "iplb_failover_ip_detail_loading_error": "Erreur lors du chargement des IPs Failover.",
+  "iplb_nat_ip_detail_loading_error": "Erreur lors du chargement des IPv4s de sortie.",
+  "iplb_status_loading_error": "Erreur lors du chargement de la section Statut.",
+  "iplb_task_list_loading_error": "Erreur lors du chargement des tâches.",
+  "iplb_farm_list_loading_error": "Erreur lors du chargement des fermes de serveurs.",
+  "iplb_farm_server_list_loading_error": "Erreur lors du chargement des serveurs de la ferme.",
+  "iplb_ipfo_select_placeholder": "Aucune",
+  "iplb_zone_select_placeholder": "Sélectionner",
+  "iplb_zone_all": "Tous (ALL)",
+  "iplb_graph_name_conn": "Connexions simultanées",
+  "iplb_graph_name_reqm": "Requêtes par minute",
+  "iplb_graph_no_data": "Aucune donnée pour le moment",
+  "iplb_status_active": "{{activeCount}} actifs",
+  "iplb_status_success_service": "Ok",
+  "iplb_status_error_service": "Erreur",
+  "iplb_status_warning_service": "Avertissement",
+  "iplb_status_help_service": "Aucune information",
+  "iplb_status_success_frontends": "{{activeCount}} sur {{totalCount}} actifs",
+  "iplb_status_error_frontends": "{{activeCount}} sur {{totalCount}} en erreur",
+  "iplb_status_warning_frontends": "{{activeCount}} sur {{totalCount}} en avertissement",
+  "iplb_status_help_frontends": "{{activeCount}} sur {{totalCount}} sans information",
+  "iplb_status_success_farms": "{{activeCount}} sur {{totalCount}} fonctionnelles",
+  "iplb_status_error_farms": "{{activeCount}} sur {{totalCount}} en erreur",
+  "iplb_status_warning_farms": "{{activeCount}} sur {{totalCount}} en avertissement",
+  "iplb_status_help_farms": "{{activeCount}} sur {{totalCount}} sans information",
+  "iplb_status_success_servers": "{{activeCount}} sur {{totalCount}} actifs",
+  "iplb_status_error_servers": "{{activeCount}} sur {{totalCount}} en erreur",
+  "iplb_status_warning_servers": "{{activeCount}} sur {{totalCount}} en avertissement",
+  "iplb_status_help_servers": "{{activeCount}} sur {{totalCount}} sans information",
+  "iplb_status_apply": "Appliquer la configuration",
+  "iplb_status_details": "Aller au détail",
+  "iplb_utilisation_update_alert": "Modifier le seuil d'alerte",
+  "iplb_utilisation_update_alert_included_data": "Données incluses",
+  "iplb_utilisation_update_alert_threshold": "Seuil de l'alerte",
+  "iplb_utilisation_update_alert_error": "Une erreur est survenue lors de la mise à jour du seuil d'alerte.",
+  "iplb_usage_loading_error": "Erreur lors du chargement de la section Utilisation.",
+  "iplb_gigabytes": "Go",
+  "iplb_cancel": "Annuler",
+  "iplb_activate": "Activer",
+  "iplb_deactivate": "Désactiver",
+  "iplb_edit": "Modifier",
+  "iplb_consult": "Consulter",
+  "iplb_manage": "Gérer",
+  "iplb_modify": "Modifier",
+  "iplb_select": "Sélectionner",
+  "iplb_add": "Ajouter",
+  "iplb_delete": "Supprimer",
+  "iplb_remove": "Retirer",
+  "iplb_modal_ok": "OK",
+  "iplb_tile_subscription": "Abonnement",
+  "iplb_tile_subscription_offer": "Offre",
+  "iplb_tile_subscription_contact": "Contacts",
+  "iplb_tile_subscription_contact_admin": "Administrateur",
+  "iplb_tile_subscription_contact_billing": "Facturation",
+  "iplb_tile_subscription_contact_technical": "Technique",
+  "iplb_tile_subscription_creation_date": "Date de création",
+  "iplb_preview_see": "Voir l'aperçu",
+  "iplb_complete_order": "Régler le bon de commande",
+  "iplb_tile_subscription_autorenew": "Renouvellement automatique"
+}
