--- packages/manager/modules/iplb/src/iplb/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "iplb_title": "IP LoadBalancer"
-}
