--- packages/manager/modules/iplb/src/template.html
@@ -1,3 +1,0 @@
-<div class="cui-page iplb">
-    <div data-ui-view="iplbContainer"></div>
-</div>
